+++ src/index.ts
@@ -23,6 +23,7 @@ export { PublishSubscribeService } from './PublishSubscribe';
 export { FlowService } from './Flow';
 export { GeneralInformationService } from './GeneralInformation';
 export { NMB2BError } from './utils/NMB2BError';
+
 export interface B2BClient {
   Airspace: AirspaceService;
   Flight: FlightService;