+++ src/utils/internals.ts
@@ -1,0 +1,5 @@
+import injectSendTime_ from './injectSendTime';
+import responseStatusHandler_ from './responseStatusHandler';
+
+export const injectSendTime = injectSendTime_;
+export const responseStatusHandler = responseStatusHandler_;